Change an argument type to compile against BoringSSL headers.
For some reason, there's a difference between how OpenSSL and BoringSSL
specify the type of this callback argument ('bio_info_cb *' vs
'bio_info_cb', respectively). While these should generally be
exchangeable, it turns out that using the former type with the latter
headers declarations fails due to an extra dereferencing level.

This fixes the problem with BoringSSL and is backward compatible when
building against OpenSSL.

Bug: 22373707
Change-Id: Ibad02749fc46f7be23d3e5b4fa205da130549c2b