tlsdate: prevent unnecessary rebuilds
tlsdate rebuilds every time the makefiles are reparsed because it embeds
the output of a shell date command into the command line, and ninja
reruns the command every time it changes.  Replace $(shell date +%sL)
with $$$(BUILD_NUMBER_FROM_FILE)L, which will read the build date from
$(OUT_DIR)/build_date.txt.  This date file will be updated on every
build, but will not cause a rebuild if it changes, so the date from the
most recent time the file was recompiled for other reasons will be used.

Bug: 24790431
Change-Id: Ib93ead6512706e13ae6dbd271e7560a3e8c00e8a